requirements.txt
#ไลบรารีสำหรับประมวลผลการแปลง PDF เป็นภาพ (PyMuपनाPDF)

PyMuPDF==1.27.1

#ไลบรารีสำหรับจัดการและนำภาพมาต่อกัน

Pillow==12.1.1

#ไลบรารีสำหรับสร้างและจัดการไฟล์ Word (.docx)

python-docx==1.2.0

# GUI framework
customtkinter==5.2.2
